==> Scripts/6a_Method_AC.ipynb <==
# Data Wrangling Plenar Protokols

##https://medium.com/@bedigunjit/simple-guide-to-text-classification-nlp-using-svm-and-naive-bayes-with-python-421db3a72d34

import pandas as pd
import nltk 
import numpy as np
import sklearn.datasets
from nltk.tokenize import word_tokenize
from nltk import pos_tag
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from sklearn.preprocessing import LabelEncoder
from collections import defaultdict
from nltk.corpus import wordnet as wn
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn import model_selection, naive_bayes, svm
from sklearn.metrics import accuracy_score
pd.set_option('display.max_columns', None)
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer
from nltk.corpus import stopwords
nltk.download('stopwords')
german_stop_words = stopwords.words('german')

df_test = pd.read_csv('/Users/lukas/OneDrive - University of Warwick/Dokumente/Uni/Thesis/Thesis/data/df_base.csv')
df_train = pd.read_csv('/Users/lukas/OneDrive - University of Warwick/Dokumente/Uni/Thesis/Thesis/data/df_train.csv')

df_test = df_test.sample(frac = 1)
df_train = df_train.sample(frac = 1)

print(len(df_test))
print(len(df_train))

X_train = df_train['speechContent']
y_train = df_train['Ruling_Party']
X_test = df_test['speechContent']
Y_test = df_test['Ruling_Party']

count_vect = CountVectorizer(stop_words = german_stop_words)

# Fit and tranform with X_train
count_vector = sklearn.feature_extraction.text.CountVectorizer(stop_words = german_stop_words)
word_counts = count_vector.fit_transform(X_train)
tf_transformer = sklearn.feature_extraction.text.TfidfTransformer(use_idf=True)
X_train = tf_transformer.fit_transform(word_counts)

SMV = svm.SVC(C=1.0, kernel='linear', degree=3, gamma='auto', probability=True)
SGD = sklearn.linear_model.SGDClassifier(loss = 'log', penalty='l2')

SMV.fit(X_train, y_train)
SGD.fit(X_train, y_train)

# Transform X_test
test_word_counts = count_vector.transform(X_test) 
ready_to_be_predicted = tf_transformer.transform(test_word_counts)

Transforming Term Frequency matrix

# predict the labels on validation dataset
predictions_SVM = SMV.predict(ready_to_be_predicted)
predictions_SGD = SGD.predict(ready_to_be_predicted)
# Use accuracy_score function to get tghe accuracy
print("SVM Accuracy Score -> ",accuracy_score(predictions_SVM, Y_test)*100)
print("SGD Accuracy Score -> ",accuracy_score(predictions_SGD, Y_test)*100)

class_probabilities_SMV = SMV.predict_proba(ready_to_be_predicted)
class_probabilities_SGD = SGD.predict_proba(ready_to_be_predicted)

confidence_0_SMV = class_probabilities_SMV.transpose()[0]
confidence_1_SMV = class_probabilities_SMV.transpose()[1]
confidence_0_SGD = class_probabilities_SGD.transpose()[0]
confidence_1_SGD = class_probabilities_SGD.transpose()[1]

df_test['SMV_predictions'] = predictions_SVM
df_test['SGD_predictions'] = predictions_SGD
df_test['Confidence_0_SMV'] = confidence_0_SMV
df_test['Confidence_1_SMV'] = confidence_1_SMV


df_test.to_csv('/Users/lukas/OneDrive - University of Warwick/Dokumente/Uni/Thesis/Thesis/data/df_base_ML.csv')

df_test
